% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpingProcedure.R
\name{clumpingProcedure}
\alias{clumpingProcedure}
\title{Clumping procedure for SLOPE}
\usage{
clumpingProcedure(screenResult, rho = 0.5, verbose = TRUE)
}
\arguments{
\item{screenResult}{object of class screenResult}

\item{rho}{numeric, minimal correlation between two SNPs to be assigned to one clump}

\item{verbose}{logical, if TRUE (default) progress bar is shown}
}
\value{
object of class \code{\link{clumpingResult}}
}
\description{
Clumping procedure for SNPs previously read and screened by
function \code{\link{readSNPs}}. Highly correlated SNPs are put in one cluster.
}

