% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpingResult.R
\name{clumpingResult}
\alias{clumpingResult}
\title{clumpingResult class}
\description{
A result of procedure for snp clumping produced by \code{clumpProcedure}
}
\details{
Always a named list of ten elements
\enumerate{
\item \code{X} numeric matrix, contains snp representative for each clump
\item \code{y} numeric vector, phenotype
\item \code{SNPnumber} numeric vector, which columns in SNP matrix \code{X_all}
are related to clumps representatives
\item \code{SNPclumps} list of numeric vectors, which columns in SNP matrix
\code{X_all} are related to clump members
\item \code{X_info} data.frame, mapping information about SNPs from .map file.
Copied from the result of screening procedure.
\item \code{selectedSnpsNumbers} numeric vector, which rows of \code{X_info}
matrix are related to selected clump representatives
\item \code{X_all} X matrix that was input for clumping procedure
\item \code{numberOfSnps} Number of SNPs in input for screening procedure
\item \code{selectedSnpsNumbersScreening} Contains information on which rows of
\code{X_info} matrix refer to SNPs that are input in clumping procedure
\item \code{pVals} p-values from marginal tests
\item \code{pValMax} p-value used in screening procedure
}
}
\seealso{
\code{\link{screeningResult}} \code{\link{clumpProcedure}}
}

