% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpProcedure.R
\name{clumpProcedure}
\alias{clumpProcedure}
\title{Clumping procedure}
\usage{
clumpProcedure(y, X, rho = 0.3, pValMax = 0.1)
}
\arguments{
\item{y}{numeric vector, phenotype}

\item{X}{numeric matrix, SNPs}

\item{rho}{numeric, minimal correlation between two SNPs to be
classified to the same clump}
}
\description{
clumping of SNPs as preprocesing for SLOPE in GWAS
}

