% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cps.R
\docType{package}
\name{cps}
\alias{cps}
\alias{cps-package}
\title{Genome Wide Association Study with SLOPE}
\description{
Genome-wide association study performed with SLOPE.
In first step, SNPs are clumped according to their correlations and
distances. Then, SLOPE is performed on data where each clump has
one representative.
}
\details{
Version: 0.27.0
}
\examples{
\donttest{
clumpedSLOPE(y, SNPs)
}
\donttest{
runExample()
}
}
\author{
Piotr Sobczyk

Maintainer: Piotr Sobczyk \email{Piotr.Sobczyk@pwr.edu.pl}
}

