% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genSLOPE.R
\name{genSLOPE}
\alias{genSLOPE}
\title{GWAS with SLOPE}
\usage{
genSLOPE(clumpingResult, fdr = 0.1, lambda = "gaussian", verbose = TRUE)
}
\arguments{
\item{clumpingResult}{clumpProcedure output}

\item{fdr,}{False Discovery Rate for SLOPE}

\item{lambda}{lambda for SLOPE. See \code{\link[SLOPE]{create_lambda}}}

\item{verbose}{if TRUE progress bar is printed}
}
\value{
object of class \code{\link{genSlopeResult}}
}
\description{
Performs GWAS with SLOPE on given snp matrix and phenotype.
At first clumping procedure is performed. Highly correlated
(that is stronger than parameter \emph{rho}) snps are clustered.
Then SLOPE is used on snp matrix which contains
one representative for each clump.
}
\examples{
\dontrun{
slope.result <- genSLOPE(clumping, fdr=0.1)
}
}

