% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_clump.R
\name{identify_clump.clumpingResult}
\alias{identify_clump.clumpingResult}
\title{Identify clump number in clumpingResult class plot}
\usage{
\method{identify_clump}{clumpingResult}(x, ...)
}
\arguments{
\item{x}{clumpingResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Identify clump number in clumpingResult class plot
}
