% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identify_clump.R
\name{identify_clump.geneSlopeResult}
\alias{identify_clump.geneSlopeResult}
\title{Identify clump number in geneSlopeResult class plot}
\usage{
\method{identify_clump}{geneSlopeResult}(x, ...)
}
\arguments{
\item{x}{geneSlopeResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Identify clump number in geneSlopeResult class plot
}

