% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneSlopeResult.R
\name{identify_clump}
\alias{identify_clump}
\title{Identify clump number in geneSlopeResult class plot}
\usage{
identify_clump(x, ...)
}
\arguments{
\item{x}{geneSlopeResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Identify clump number in geneSlopeResult class plot
}

