% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{GWAS with clumped SLOPE}
\usage{
main(phenotypeFile = NULL, snpFiles = NULL, pValMax = 0.05, rho = 0.5,
  fdr = 0.1, header = T, sep = ",", verbose = TRUE)
}
\arguments{
\item{phenotypeFile}{file containing phentotype as its first column}

\item{snpFiles}{vector of filenames with SNPs e.g. different chromosomes}

\item{pValMax}{threshold p-value in marginal snp test}

\item{rho}{threshold value of correlation for snp clumping}

\item{fdr}{False Discovery Rate for SLOPE procedure}

\item{header}{logical, whether file with phenotype contains header}

\item{sep}{seperator for file with phenotype}

\item{verbose}{if TRUE (default) progress bar and info messages shall
be displayed}
}
\description{
Function performes Genome Wide Association Study (GWAS) using SLOPE.
In first step, user provides files with phenotype and snps as either
vector of strings or interactively via dialog window.
In second step, data is read and snps are initially screened using
threshold p-value provided by the user.
In third step, clumping procedure is performed. Highly correlated
(that is stronger than parameter \emph{rho}) snps are clustered.
In last step, SLOPE is performed on representative of each clump.
}
\details{
Files with SNPs should be of .raw file format
}

