% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeData.R
\name{phenotypeData}
\alias{phenotypeData}
\title{phenotypeData class}
\description{
Phenotype data
}
\details{
Always a named list of two elements
\enumerate{
\item \code{y} numeric vector, phenotype
\item \code{yInfo} data.frame, additional information about observations
provied in .fam file
}
}
\seealso{
\code{\link{read_phenotype}}
}
