% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpProcedure.R
\name{plot.clumpingResult}
\alias{plot.clumpingResult}
\title{Plot clumpingResult class object}
\usage{
\method{plot}{clumpingResult}(x, chromosomeNumber = NULL, ...)
}
\arguments{
\item{x}{clumpingResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}

\item{chromosome}{optional parameter, only selected chromosome will be plotted}
}
\description{
Plot clumpingResult class object
}
\keyword{internal}

