% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneSlopeResult.R
\name{plot.geneSlopeResult}
\alias{plot.geneSlopeResult}
\title{Plot geneSlopeResult class object}
\usage{
\method{plot}{geneSlopeResult}(x, chromosomeNumber = NULL,
  clumpNumber = NULL, ...)
}
\arguments{
\item{x}{geneSlopeResult class object}

\item{chromosomeNumber}{optional parameter, only selected chromosome will be plotted}

\item{clumpNumber}{optional parameter, only SNPs from selected clump will be plotted}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Plot geneSlopeResult class object
}

