% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genSLOPE.R
\name{plot.genSlopeResult}
\alias{plot.genSlopeResult}
\title{Plot genSlopeResult class object}
\usage{
\method{plot}{genSlopeResult}(x, chromosomeNumber = NULL, ...)
}
\arguments{
\item{x}{genSlopeResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}

\item{chromosome}{optional parameter, only selected chromosome will be plotted}
}
\description{
Plot genSlopeResult class object
}
\keyword{internal}

