% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionResult.R
\name{plot.selectionResult}
\alias{plot.selectionResult}
\title{Plot selectionResult class object}
\usage{
\method{plot}{selectionResult}(x, chromosomeNumber = NULL, clumpNumber = NULL, ...)
}
\arguments{
\item{x}{selectionResult class object}

\item{chromosomeNumber}{optional parameter, only selected chromosome will be plotted}

\item{clumpNumber}{optional parameter, only SNPs from selected clump will be plotted}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Plot selectionResult class object
}
