% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpingResult.R
\name{print.clumpingResult}
\alias{print.clumpingResult}
\title{Print clumpingResult class object}
\usage{
\method{print}{clumpingResult}(x, ...)
}
\arguments{
\item{x}{clumpingResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Print clumpingResult class object
}
