% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneSlopeResult.R
\name{print.geneSlopeResult}
\alias{print.geneSlopeResult}
\title{Print geneSlopeResult class object}
\usage{
\method{print}{geneSlopeResult}(x, ...)
}
\arguments{
\item{x}{geneSlopeResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Nothing.
}
\description{
Print geneSlopeResult class object
}

