% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeData.R
\name{print.phenotypeData}
\alias{print.phenotypeData}
\title{Print phenotypeData class object}
\usage{
\method{print}{phenotypeData}(x, ...)
}
\arguments{
\item{x}{phenotypeData class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Print phenotypeData class object
}
