% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSNPs.R
\name{print.screeningResult}
\alias{print.screeningResult}
\title{Print screeningResult class object}
\usage{
\method{print}{screeningResult}(x, ...)
}
\arguments{
\item{x}{screeningResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Print screeningResult class object
}
\keyword{internal}

