% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionResult.R
\name{print.selectionResult}
\alias{print.selectionResult}
\title{Print selectionResult class object}
\usage{
\method{print}{selectionResult}(x, ...)
}
\arguments{
\item{x}{selectionResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Nothing.
}
\description{
Print selectionResult class object
}
