% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readBigSNPs.R
\name{readBigSNPs}
\alias{readBigSNPs}
\title{Reading SNPs from one PLINK .raw and .map files}
\usage{
readBigSNPs(rawFile, mapFile = "", y, pValMax = 0.05, chunk_size = 1000,
  verbose = TRUE)
}
\arguments{
\item{rawFile}{name of .raw file}

\item{mapFile}{name of .map file}

\item{y}{phenotype}

\item{pValMax}{p-value threshold value used for screening}

\item{verbose}{if TRUE (default) information about progress is printed}

\item{chunk.size}{size of chunk. The bigger the chunk the faster function works but
computer might run out of RAM}
}
\value{
object of class \code{\link{screeningResult}}
}
\description{
Reading PLINK files that were previously suitably tranformed
in plink - see details
}
\details{
\strong{Exporting data from PLINK}
To import data to R, it needs to be exported from
PLINK using the option "--recodeAD"
The PLINK command should therefore look like
\code{plink --file input --recodeAD --out output}.
For more information, please refer to:
\url{http://pngu.mgh.harvard.edu/~purcell/plink/dataman.shtml}
}

