% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readPLINK.R
\name{readPLINK}
\alias{readPLINK}
\title{Reading PLINK .raw files}
\usage{
readPLINK(file)
}
\arguments{
\item{file}{name of .raw file}
}
\value{
list \itemize{
\item snps matrix of SNPs
\item phenotype vector with phenotype values
\item positions vector with information about SNP positions in chromosome
}
}
\description{
Reading PLINK files that were previously suitably tranformed
in plink - see details
}
\details{
\strong{Exporting data from PLINK}
As stated in documentation for function \code{\link[adegenet]{read.PLINK}},
to import data to R, it needs to be exported from
PLINK using the option "--recodeA" (and NOT "--recodeAD").
The PLINK command should therefore look like:
\code{plink --file data --recodeA}.
For more information on this topic, please look at this webpage:
\url{http://pngu.mgh.harvard.edu/~purcell/plink/dataman.shtml}
}

