% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SNPs}
\alias{SNPs}
\title{Matrix containing SNPs}
\format{Large numeric matrix, 75510 elements}
\usage{
SNPs
}
\description{
Subset of PLINK example data, available at PLINK
website
}
\keyword{datasets}

