% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpProcedure.R
\name{summary.clumpingResult}
\alias{summary.clumpingResult}
\title{Summary clumpingResult class object}
\usage{
\method{summary}{clumpingResult}(x, ...)
}
\arguments{
\item{x}{clumpingResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Summary clumpingResult class object
}
\keyword{internal}

