% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genSlopeResult.R
\name{summary.genSlopeResult}
\alias{summary.genSlopeResult}
\title{Summary genSlopeResult class object}
\usage{
\method{summary}{genSlopeResult}(x, ...)
}
\arguments{
\item{x}{genSlopeResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Summary genSlopeResult class object
}

