% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeData.R
\name{summary.phenotypeData}
\alias{summary.phenotypeData}
\title{Summary phenotypeData class object}
\usage{
\method{summary}{phenotypeData}(object, ...)
}
\arguments{
\item{object}{phenotypeData class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Summary phenotypeData class object
}
