% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSNPs.R
\name{summary.screeningResult}
\alias{summary.screeningResult}
\title{Summary screeningResult class object}
\usage{
\method{summary}{screeningResult}(x, ...)
}
\arguments{
\item{x}{screeningResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Summary screeningResult class object
}
\keyword{internal}

