% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionResult.R
\name{summary.selectionResult}
\alias{summary.selectionResult}
\title{Summary selectionResult class object}
\usage{
\method{summary}{selectionResult}(object, clumpNumber = NULL, ...)
}
\arguments{
\item{object}{selectionResult class object}

\item{clumpNumber}{number of clump to be summarized}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Summary selectionResult class object
}
