% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformPedToRaw.R
\name{transformPedToRaw}
\alias{transformPedToRaw}
\title{Transformin .ped to .raw using plink command line}
\usage{
transformPedToRaw(file)
}
\arguments{
\item{file,}{character, name of .raw file}
}
\description{
Transformin .ped to .raw using plink command line
}

