% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{y}
\alias{y}
\title{Phenotype}
\format{a numeric vector of length 90. Each element refers
to phenotype for one observations and is related to one
row in \code{\link{SNPs}} matrix}
\usage{
aaaggregation
}
\description{
Artificial phenotype. Created as a linear combination
of small number of columns of \code{\link{SNPs}} matrix plus some noise
}
\keyword{datasets}

