% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpedSLOPE.R
\name{clumpedSLOPE}
\alias{clumpedSLOPE}
\title{Clumped SLOPE}
\usage{
clumpedSLOPE(y, X, rho = 0.3, fdr = 0.2)
}
\arguments{
\item{y}{numeric vector, phenotype}

\item{X}{numeric matrix, SNPs}

\item{rho}{numeric, minimal correlation between two SNPs to be
classified to the same clump}

\item{fdr,}{False Discovery Rate for SLOPE}
}
\description{
Clumped SLOPE
}

