% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpingProcedure.R
\name{clumpingProcedure}
\alias{clumpingProcedure}
\title{Clumping procedure for SLOPE}
\usage{
clumpingProcedure(screenResult, rho = 0.5, verbose = TRUE)
}
\arguments{
\item{screenResult}{object of class screeningResult}

\item{rho}{numeric, minimal correlation between two SNPs to be assigned to one clump}

\item{verbose}{logical, if TRUE (default) progress bar is shown}
}
\value{
object of class \code{\link{clumpingResult}}
}
\description{
Clumping procedure performed on SNPs, columns of matrix \code{X}, from
object of class \code{\link{screeningResult}},
which is an output of function \code{\link{readSNPs}}.
SNPs are clustered based on their correlations. For details see package vignette.
}

