% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpingResult.R
\name{clumpingResult}
\alias{clumpingResult}
\title{clumpingResult class}
\description{
A result of procedure for snp clumping produced by \code{\link{clump_snps}}
}
\details{
Always a named list of eleven elements
\enumerate{
\item \code{X} numeric matrix, consists of one snp representative for each clump
\item \code{y} numeric vector, phenotype
\item \code{SNPnumber} numeric vector, which columns in SNP matrix \code{X_all}
are related to clumps representatives
\item \code{SNPclumps} list of numeric vectors, which columns in SNP matrix
\code{X_all} are related to clump members
\item \code{X_info} data.frame, mapping information about SNPs from .map file.
Copied from the result of screening procedure.
\item \code{selectedSnpsNumbers} numeric vector, which rows of \code{X_info}
matrix are related to selected clump representatives
\item \code{X_all} numeric matrix, all the snps that passed screening procedure
\item \code{numberOfSnps} numeric, total number of SNPs before screening procedure
\item \code{selectedSnpsNumbersScreening} numeric vector, which rows of \code{X_info}
data.frame are related to snps that passed screening
\item \code{pVals} numeric vector, p-values from marginal tests for each snp
\item \code{pValMax} numeric, p-value used in screening procedure
}
}
\seealso{
\code{\link{screeningResult}} \code{\link{clump_snps}}
}
