% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpProcedure.R
\name{clumpProcedure}
\alias{clumpProcedure}
\title{Clumping procedure}
\usage{
clumpProcedure(screenResult, rho = 0.3, verbose = TRUE)
}
\arguments{
\item{screenResult}{object of class screenResult}

\item{rho}{numeric, minimal correlation between two SNPs to be}

\item{verbose}{logical, if TRUE (default) progress bar is shown
classified to the same clump}
}
\value{
object of class \code{\link{clumpingResult}}
}
\description{
Clumping of SNPs previously read and screened by \code{\link{readBigSNPs}}
}

