% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpProcedure.R
\name{clumpProcedure}
\alias{clumpProcedure}
\title{Clumping procedure}
\usage{
clumpProcedure(y, X, rho = 0.3, verbose = TRUE)
}
\arguments{
\item{y}{numeric vector, phenotype}

\item{X}{numeric matrix, SNPs}

\item{rho}{numeric, minimal correlation between two SNPs to be}

\item{verbose}{logical, if TRUE (default) progress bar is shown
classified to the same clump}
}
\description{
Clumping of SNPs. Can be used as a preprocesing tool for using SLOPE in GWAS
}

