% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clumpedSLOPE.R
\name{genSLOPE}
\alias{genSLOPE}
\title{GWAS with SLOPE}
\usage{
genSLOPE(clumpingResult, fdr = 0.1, lambda = "gaussian", verbose = TRUE)
}
\arguments{
\item{fdr,}{False Discovery Rate for SLOPE}

\item{lambda}{lambda for SLOPE. See \code{\link[SLOPE]{create_lambda}}}

\item{clumpProcedure}{}
}
\value{
data.frame with two columns \itemize{
\item snpName names of selected snps
\item snpNumber column numbers of selected snps in input matrix X
}
}
\description{
Performs GWAS with SLOPE on given snp matrix and phenotype.
At first clumping procedure is performed. Highly correlated
(that is stronger than parameter \emph{rho}) snps are clustered.
Then SLOPE is used on snp matrix which contains
one representative for each clump.
}
\examples{
\donttest{
clumpedSLOPE(y, SNPs)
}
}

