% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_geneSLOPE.R
\name{gui_geneSLOPE}
\alias{gui_geneSLOPE}
\title{GUI for GWAS with SLOPE}
\usage{
gui_geneSLOPE()
}
\value{
No return value, called for side effects
}
\description{
A graphical user interface for performing Genome-wide
Association Study with SLOPE
}
\details{
requires installing \pkg{\link[shiny]{shiny}} package
}
