% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_clump.R
\name{identify_clump}
\alias{identify_clump}
\title{identify_clump}
\usage{
identify_clump(x, ...)
}
\arguments{
\item{x}{appropiate class object}

\item{...}{other arguments}
}
\description{
identify_clump
}
\details{
Enable interactive selection of snps in plot.
Return clump number.
}
