% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_clump.R
\name{identify_clump.selectionResult}
\alias{identify_clump.selectionResult}
\title{Identify clump number in selectionResult class plot}
\usage{
\method{identify_clump}{selectionResult}(x, ...)
}
\arguments{
\item{x}{selectionResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Identify clump number in selectionResult class plot
}
