% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{GWAS with clumped SLOPE}
\usage{
main(phenotypeFile = NULL, snpFiles = NULL, pValMax = 0.1, header = T,
  sep = ",", verbose = TRUE)
}
\arguments{
\item{phenotypeFile}{file containing phentotype as its first column}

\item{snpFiles}{vector of filenames with SNPs e.g. different chromosomes}

\item{pValMax}{threshold p-value in marginal snp test}

\item{header}{logical, whether file with phenotype contains header}

\item{sep}{seperator for file with phenotype}

\item{verbose}{if TRUE (default) progress bar and info messages shall
be displayed}
}
\description{
Files with phenotype and snps are selected interactively.
}
\details{
Files with SNPs should be of .raw file format
}

