% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpingResult.R
\name{plot.clumpingResult}
\alias{plot.clumpingResult}
\title{Plot clumpingResult class object}
\usage{
\method{plot}{clumpingResult}(x, chromosomeNumber = NULL, clumpNumber = NULL, ...)
}
\arguments{
\item{x}{clumpingResult class object}

\item{chromosomeNumber}{optional parameter, only selected chromosome will be plotted}

\item{clumpNumber}{optional parameter, only SNPs from selected clump will be plotted}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Plot clumpingResult class object
}
\keyword{internal}
