% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genSlopeResult.R
\name{print.genSlopeResult}
\alias{print.genSlopeResult}
\title{Print genSlopeResult class object}
\usage{
\method{print}{genSlopeResult}(x, ...)
}
\arguments{
\item{x}{genSlopeResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Nothing.
}
\description{
Print genSlopeResult class object
}

