% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningResult.R
\name{print.screeningResult}
\alias{print.screeningResult}
\title{Print function for class screeningResult class}
\usage{
\method{print}{screeningResult}(x, ...)
}
\arguments{
\item{x}{\code{\link{screeningResult}} class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
No return value, called for side effects
}
\description{
Print function for class screeningResult class
}
