% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readPLINK.R
\name{readPLINK}
\alias{readPLINK}
\title{Reading SNPs from one PLINK .raw files}
\usage{
readPLINK(file)
}
\arguments{
\item{file}{name of .raw file}
}
\value{
list \itemize{
\item snps matrix of SNPs
\item observationInfo matrix containing non-genetic information including
 sex and phenotype
\item snpInfo matrix with information about SNPs names and positions in chromosomes
}
}
\description{
Reading PLINK files that were previously suitably tranformed
in plink - see details
}
\details{
\strong{Exporting data from PLINK}
To import data to R, it needs to be exported from
PLINK using the option "--recodeA" (and NOT "--recodeAD").
The PLINK command should therefore look like
\code{plink --file data --recodeA}.
For more information, please refer to:
\url{http://pngu.mgh.harvard.edu/~purcell/plink/dataman.shtml}
}

