% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSNPs.R
\name{readSNPs}
\alias{readSNPs}
\title{Reading SNPs from one PLINK .tped and .tfam files}
\usage{
readSNPs(file, y, pValMax = 0.05, chunk.size = 50000, verbose = TRUE)
}
\arguments{
\item{file}{name of .tped file}

\item{y}{phenotype}

\item{pValMax}{p-value threshold value used for screening}

\item{chunk.size}{size of chunk. The bigger the chunk the faster function works but
computer might run out of RAM}

\item{verbose}{should info be present}
}
\value{
list \itemize{
\item X matrix of filtered SNPs
\item y phenotype
\item X_info SNPs info
\item numberOfSnps number of snps read from .tped file
\item pValMax p-value threshold value used for screening
}
}
\description{
Reading PLINK files that were previously suitably tranformed
in plink - see details
}
\details{
\strong{Exporting data from PLINK}
To import data to R, it needs to be exported from
PLINK using the option "--recode12 --transposed"
The PLINK command should therefore look like
\code{plink --file data --recode12 --transposed --out temp}.
For more information, please refer to:
\url{http://pngu.mgh.harvard.edu/~purcell/plink/dataman.shtml}
}

