% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningResult.R
\name{screeningResult}
\alias{screeningResult}
\title{screeningResult class}
\description{
A result of procedure for snp clumping produced by \code{\link{screen_snps}}
}
\details{
Always a named list of eight elements
\enumerate{
\item \code{X} numeric matrix, consists of snps that passed screening
\item \code{y} numeric vector, phenotype
\item \code{X_info} data.frame, SNP info from .map file
\item \code{pVals} numeric vector, p-values from marginal tests for each snp
\item \code{numberOfSnps} numeric, total number of SNPs in .raw file
\item \code{selectedSnpsNumbers} numeric vector, which rows of \code{X_info}
data.frame are related to snps that passed screening
\item \code{pValMax} numeric, p-value used in screening procedure
\item \code{phenotypeInfo} data.frame, additional information about observations
provied in \code{\link{phenotypeData}} object
}
}
\seealso{
\code{\link{phenotypeData}} \code{\link{screen_snps}}
}
