% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpingResult.R
\name{summary.clumpingResult}
\alias{summary.clumpingResult}
\title{Summary clumpingResult class object}
\usage{
\method{summary}{clumpingResult}(object, ...)
}
\arguments{
\item{object}{clumpingResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Summary clumpingResult class object
}
