% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningResult.R
\name{summary.screeningResult}
\alias{summary.screeningResult}
\title{Summary function for class screeningResult}
\usage{
\method{summary}{screeningResult}(object, ...)
}
\arguments{
\item{object}{screeningResult class object}

\item{...}{Further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Summary function for class screeningResult
}
