% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transformPedToRaw.R
\name{transformPedToRaw}
\alias{transformPedToRaw}
\title{Transforming .ped to .raw using PLINK executable}
\usage{
transformPedToRaw(plinkExec = NULL, file)
}
\arguments{
\item{plinkExec}{path to PLINK executable}

\item{file}{name of .ped file}
}
\description{
Performs PLINK in command line to transform .ped
into R compatible .raw file.
}

